use parent notifier
