+++ server/app/src/services/consultations/consultation-mail-notifier.service.ts
@@ -1,22 +1,20 @@
-import { EmailService } from '../email/email.service';
+import { getConnection, Between } from 'typeorm';
 import {
   ITimeInverval,
   ITimeFrame,
   IConsultation,
-  IParent,
-  IStudent,
 } from './consultation-mail-notifier.interfaces';
+import { INotifier } from './notifier.interface';
 import { ConsultationRepository } from '../../repositories/consultation.repository';
 import '../../utils/extensions/date.extentions';
-import { getConnection, Between } from 'typeorm';
 
 export class ConsultationNotifier {
   notificationsSentInPreviousRound = new Map();
   notificationsSentInCurrentRound = new Map();
 
   constructor(
-    private readonly emailService: EmailService,
     private readonly timeIntervalInMinutes: ITimeInverval,
+    private readonly notifiers: INotifier<IConsultation>[],
   ) {}
 
   public async notifyParentsAboutTheirChildrenConsultations(): Promise<void> {
@@ -68,47 +66,8 @@ export class ConsultationNotifier {
   }
 
   private sendNotifications(upcomingConsultations: IConsultation[]): void {
-    for (const consultation of upcomingConsultations) {
-      if (!consultation.students) break;
-      for (const student of consultation.students) {
-        const key = this.createKey(consultation, student);
-        if (this.isEligableForNotification(student, key))
-          this.sendMail(consultation, student, student.parents);
-        this.notificationsSentInCurrentRound.set(key, true);
-      }
+    for (const notifier of this.notifiers) {
+      notifier.sendNotifications(upcomingConsultations);
     }
-    this.notificationsSentInPreviousRound = this.notificationsSentInCurrentRound;
-    this.notificationsSentInCurrentRound = new Map();
-  }
-
-  private createKey(consultation: IConsultation, student: IStudent): string {
-    return consultation.id + student.id;
-  }
-
-  private isEligableForNotification(student: IStudent, key: string) {
-    return (
-      !this.notificationsSentInPreviousRound.has(key) &&
-      student.parents &&
-      student.parents.length > 0
-    );
-  }
-
-  private sendMail(
-    consultation: IConsultation,
-    student: IStudent,
-    parents: IParent[],
-  ) {
-    const sender = 'logity consultation reminder service';
-    const recipents = parents.map(parent => parent.email);
-    const subject = 'consultation reminder';
-    const body = this.createMailBody(consultation.datetime, student.firstName);
-    this.emailService.sendMail(sender, recipents, subject, body);
-  }
-
-  private createMailBody(
-    consultationDatetime: Date,
-    studentName: string,
-  ): string {
-    return `Your child ${studentName} is expected to come to consultations at ${consultationDatetime}`;
   }
 }
