+++ server/app/src/services/consultations/notifier.interface.ts
@@ -1,0 +1,3 @@
+export interface INotifier<T> {
+  sendNotifications(notifications: T[]): void;
+}
